chore: refactor opcode spammer with calldatacopy, reverts
