+++ yarn-project/bb-prover/src/avm_proving_tests/avm_opcode_spam.test.ts
@@ -1,22 +1,9 @@
-/**
- * Opcode Spammer Proving Test Suite
- *
- * Tests all spammable opcodes using the data-driven implementation with real AVM proving.
- * Uses smallest wire formats (_8) for maximum instruction density.
- *
- * This test is meant to be run locally for measurements. It is skipped in CI.
- */
 import { createLogger } from '@aztec/foundation/log';
-import {
-  createNestedSpamBytecodeFromConfig,
-  createSpamBytecodeFromConfig,
-  getAllSpamTestCases,
-} from '@aztec/simulator/public/avm/opcode_spammer';
 import {
   TestExecutorMetrics,
   defaultGlobals,
-  testCustomBytecode,
-  testNestedCustomBytecode,
+  getSpamConfigsPerOpcode,
+  testOpcodeSpamCase,
 } from '@aztec/simulator/public/fixtures';
 import { NativeWorldStateService } from '@aztec/world-state';
 
@@ -25,15 +12,21 @@ import path from 'path';
 
 import { AvmProvingTester } from './avm_proving_tester.js';
 
-// Get all test cases from the spammer config (hierarchical by opcode)
-const allTestCases = getAllSpamTestCases();
+// NOTE: this test is meant to be run locally for measurements or via bb-prover/bootstrap.sh.
+// Set RUN_AVM_OPCODE_SPAM=1 to enable.
+const describeOrSkip = process.env.RUN_AVM_OPCODE_SPAM ? describe : describe.skip;
 
-// Note: this test is meant to be run locally for measurements. It is skipped in CI.
-describe('AVM Opcode Spammer Proving Benchmarks', () => {
+describeOrSkip('AVM Opcode Spammer Proving Benchmarks', () => {
   const logger = createLogger('avm-opcode-spam-proving');
+
+  // Get all test cases from the spammer config (grouped by opcode)
+  const groupedSpamConfigs = getSpamConfigsPerOpcode();
 
   // Shared metrics instance for benchmark collection
   const metrics = new TestExecutorMetrics();
+  // Full proving only (no check-circuit mode)
+  let worldStateService: NativeWorldStateService;
+  let tester: AvmProvingTester;
 
   afterAll(() => {
     if (process.env.BENCH_OUTPUT) {
@@ -46,10 +39,6 @@ describe('AVM Opcode Spammer Proving Benchmarks', () => {
       logger.info(metrics.toPrettyString());
     }
   });
-
-  // Full proving only (no check-circuit mode)
-  let worldStateService: NativeWorldStateService;
-  let tester: AvmProvingTester;
 
   beforeEach(async () => {
     worldStateService = await NativeWorldStateService.tmp();
@@ -67,20 +56,11 @@ describe('AVM Opcode Spammer Proving Benchmarks', () => {
     await worldStateService.close();
   });
 
-  // Unified test loop - handles both gas-limited and side-effect-limited opcodes
-  describe.each(allTestCases)('$opcode', ({ opcode, cases }) => {
-    it.each(cases.map(c => ({ ...c, opcode })))(
-      '$variant',
-      async ({ opcode, variant, config, isNested }) => {
-        const label = variant !== opcode ? `${opcode}/${variant}` : opcode;
-        if (isNested) {
-          const { innerBytecode, createOuterBytecode } = createNestedSpamBytecodeFromConfig(config);
-          await testNestedCustomBytecode(innerBytecode, createOuterBytecode, tester, label);
-        } else {
-          const { bytecode } = createSpamBytecodeFromConfig(config);
-          await testCustomBytecode(bytecode, tester, label);
-        }
-        // No result checks - this is for proving benchmarks only
+  describe.each(groupedSpamConfigs)('$opcode', ({ configs }) => {
+    it.each(configs)(
+      '$label',
+      async config => {
+        await testOpcodeSpamCase(tester, config);
       },
       600_000,
     );
