+++ yarn-project/simulator/package.json
@@ -7,8 +7,7 @@
     "./client": "./dest/client.js",
     "./testing": "./dest/testing.js",
     "./public/fixtures": "./dest/public/fixtures/index.js",
-    "./public/avm/opcodes": "./dest/public/avm/opcodes/index.js",
-    "./public/avm/opcode_spammer": "./dest/public/avm/opcode_spammer/index.js"
+    "./public/avm/opcodes": "./dest/public/avm/opcodes/index.js"
   },
   "typedocOptions": {
     "entryPoints": [
