+++ yarn-project/simulator/src/public/fixtures/index.ts
@@ -7,4 +7,9 @@ export { ammTest } from './amm_test.js';
 export { bulkTest, megaBulkTest } from './bulk_test.js';
 export { tokenTest } from './token_test.js';
 export * from './custom_bytecode_tests.js';
-export { testCustomBytecode, testNestedCustomBytecode } from './custom_bytecode_tester.js';
+export {
+  deployCustomBytecode,
+  executeCustomBytecode,
+  deployAndExecuteCustomBytecode,
+} from './custom_bytecode_tester.js';
+export { getSpamConfigsPerOpcode, testOpcodeSpamCase } from './opcode_spammer.js';
